+++ src/app/store.ts
@@ -4,7 +4,7 @@ import {
   useDispatch as dispatchHook,
   useSelector as selectorHook,
 } from 'react-redux';
-import { pokemonsReducer } from '../entities/pokemon/model/pokemonsSlice';
+import { pokemonsReducer } from '../entities/pokemon';
 
 export const rootReducer = combineReducers({
   pokemons: pokemonsReducer,
