+++ src/widgets/ResultsList/ui/ResultList.test.tsx
@@ -4,21 +4,12 @@ import { ResultList } from './ResultList';
 import { render, screen } from '@testing-library/react';
 import { MemoryRouter } from 'react-router';
 
-describe('mock', () => {
-  test('mock', () => {
-    expect(1).toBe(1);
-  });
-});
-
-vi.mock('../../../entities/pokemon/ui/ListCard', () => ({
+vi.mock('../../../entities/pokemon', () => ({
   ListCard: ({ name, url }: { name: string; url: string }) => (
     <li data-testid="pokemon-card">
       {name} - {url}
     </li>
   ),
-}));
-
-vi.mock('../../../entities/pokemon/ui/PokemonCardSkeleton', () => ({
   PokemonSkeletonCard: () => <li data-testid="skeleton-card" />,
 }));
